DEV: Handle ArrayConstructor in ArraySize_t_util
