+++ src/libasr/pass/simplifier.cpp
@@ -65,176 +65,6 @@ class ArrayVarCollector: public ASR::BaseWalkVisitor<ArrayVarCollector> {
 
 };
 
-ASR::expr_t* get_ImpliedDoLoop_size(Allocator& al, ASR::ImpliedDoLoop_t* implied_doloop) {
-    const Location& loc = implied_doloop->base.base.loc;
-    ASRUtils::ASRBuilder builder(al, loc);
-    ASR::expr_t* start = implied_doloop->m_start;
-    ASR::expr_t* end = implied_doloop->m_end;
-    ASR::expr_t* d = implied_doloop->m_increment;
-    ASR::expr_t* implied_doloop_size = nullptr;
-    int kind = ASRUtils::extract_kind_from_ttype_t(ASRUtils::expr_type(end));
-    start = builder.i2i_t(start, ASRUtils::expr_type(end));
-    if( d == nullptr ) {
-        implied_doloop_size = builder.Add(
-            builder.Sub(end, start),
-            make_ConstantWithKind(make_IntegerConstant_t, make_Integer_t, 1, kind, loc));
-    } else {
-        implied_doloop_size = builder.Add(builder.Div(
-            builder.Sub(end, start), d),
-            make_ConstantWithKind(make_IntegerConstant_t, make_Integer_t, 1, kind, loc));
-    }
-    int const_elements = 0;
-    ASR::expr_t* implied_doloop_size_ = nullptr;
-    for( size_t i = 0; i < implied_doloop->n_values; i++ ) {
-        if( ASR::is_a<ASR::ImpliedDoLoop_t>(*implied_doloop->m_values[i]) ) {
-            if( implied_doloop_size_ == nullptr ) {
-                implied_doloop_size_ = get_ImpliedDoLoop_size(al,
-                    ASR::down_cast<ASR::ImpliedDoLoop_t>(implied_doloop->m_values[i]));
-            } else {
-                implied_doloop_size_ = builder.Add(get_ImpliedDoLoop_size(al,
-                    ASR::down_cast<ASR::ImpliedDoLoop_t>(implied_doloop->m_values[i])),
-                    implied_doloop_size_);
-            }
-        } else {
-            const_elements += 1;
-        }
-    }
-    if( const_elements > 1 ) {
-        if( implied_doloop_size_ == nullptr ) {
-            implied_doloop_size_ = make_ConstantWithKind(make_IntegerConstant_t,
-                make_Integer_t, const_elements, kind, loc);
-        } else {
-            implied_doloop_size_ = builder.Add(
-                make_ConstantWithKind(make_IntegerConstant_t,
-                    make_Integer_t, const_elements, kind, loc),
-                implied_doloop_size_);
-        }
-    }
-    if( implied_doloop_size_ ) {
-        implied_doloop_size = builder.Mul(implied_doloop_size_, implied_doloop_size);
-    }
-    return implied_doloop_size;
-}
-
-size_t get_constant_ArrayConstant_size(ASR::ArrayConstant_t* x) {
-    return ASRUtils::get_fixed_size_of_array(x->m_type);
-}
-
-ASR::expr_t* get_ArrayConstant_size(Allocator& al, ASR::ArrayConstant_t* x) {
-    ASR::ttype_t* int_type = ASRUtils::TYPE(ASR::make_Integer_t(al, x->base.base.loc, 4));
-    return make_ConstantWithType(make_IntegerConstant_t,
-            ASRUtils::get_fixed_size_of_array(x->m_type), int_type, x->base.base.loc);
-}
-
-ASR::expr_t* get_ArrayConstructor_size(Allocator& al, ASR::ArrayConstructor_t* x) {
-    ASR::ttype_t* int_type = ASRUtils::TYPE(ASR::make_Integer_t(al, x->base.base.loc, 4));
-    ASR::expr_t* array_size = nullptr;
-    int64_t constant_size = 0;
-    const Location& loc = x->base.base.loc;
-    ASRUtils::ASRBuilder builder(al, loc);
-    for( size_t i = 0; i < x->n_args; i++ ) {
-        ASR::expr_t* element = x->m_args[i];
-        if( ASR::is_a<ASR::ArrayConstant_t>(*element) ) {
-            if( ASRUtils::is_value_constant(element) ) {
-                constant_size += get_constant_ArrayConstant_size(
-                    ASR::down_cast<ASR::ArrayConstant_t>(element));
-            } else {
-                ASR::expr_t* element_array_size = get_ArrayConstant_size(al,
-                    ASR::down_cast<ASR::ArrayConstant_t>(element));
-                if( array_size == nullptr ) {
-                    array_size = element_array_size;
-                } else {
-                    array_size = builder.Add(array_size,
-                                    element_array_size);
-                }
-            }
-        } else if( ASR::is_a<ASR::ArrayConstructor_t>(*element) ) {
-            ASR::expr_t* element_array_size = get_ArrayConstructor_size(al,
-                ASR::down_cast<ASR::ArrayConstructor_t>(element));
-            if( array_size == nullptr ) {
-                array_size = element_array_size;
-            } else {
-                array_size = builder.Add(array_size,
-                                element_array_size);
-            }
-        } else if( ASR::is_a<ASR::Var_t>(*element) ) {
-            ASR::ttype_t* element_type = ASRUtils::type_get_past_allocatable(
-                ASRUtils::expr_type(element));
-            if( ASRUtils::is_array(element_type) ) {
-                if( ASRUtils::is_fixed_size_array(element_type) ) {
-                    ASR::dimension_t* m_dims = nullptr;
-                    size_t n_dims = ASRUtils::extract_dimensions_from_ttype(element_type, m_dims);
-                    constant_size += ASRUtils::get_fixed_size_of_array(m_dims, n_dims);
-                } else {
-                    ASR::expr_t* element_array_size = ASRUtils::get_size(element, al);
-                    if( array_size == nullptr ) {
-                        array_size = element_array_size;
-                    } else {
-                        array_size = builder.Add(array_size,
-                                        element_array_size);
-                    }
-                }
-            } else {
-                constant_size += 1;
-            }
-        } else if( ASR::is_a<ASR::ImpliedDoLoop_t>(*element) ) {
-            ASR::expr_t* implied_doloop_size = get_ImpliedDoLoop_size(al,
-                ASR::down_cast<ASR::ImpliedDoLoop_t>(element));
-            if( array_size ) {
-                array_size = builder.Add(implied_doloop_size, array_size);
-            } else {
-                array_size = implied_doloop_size;
-            }
-        } else if( ASR::is_a<ASR::ArraySection_t>(*element) ) {
-            ASR::ArraySection_t* array_section_t = ASR::down_cast<ASR::ArraySection_t>(element);
-            ASR::expr_t* array_section_size = nullptr;
-            for( size_t j = 0; j < array_section_t->n_args; j++ ) {
-                ASR::expr_t* start = array_section_t->m_args[j].m_left;
-                ASR::expr_t* end = array_section_t->m_args[j].m_right;
-                ASR::expr_t* d = array_section_t->m_args[j].m_step;
-                if( d == nullptr ) {
-                    continue;
-                }
-                ASR::expr_t* dim_size = builder.Add(builder.Div(
-                    builder.Sub(end, start), d),
-                    make_ConstantWithKind(make_IntegerConstant_t, make_Integer_t, 1, 4, loc));
-                if( array_section_size == nullptr ) {
-                    array_section_size = dim_size;
-                } else {
-                    array_section_size = builder.Mul(array_section_size, dim_size);
-                }
-            }
-            if( array_size == nullptr ) {
-                array_size = array_section_size;
-            } else {
-                builder.Add(array_section_size, array_size);
-            }
-        } else {
-            constant_size += 1;
-        }
-    }
-    ASR::expr_t* constant_size_asr = nullptr;
-    if (constant_size == 0 && array_size == nullptr) {
-        constant_size = ASRUtils::get_fixed_size_of_array(x->m_type);
-    }
-    if( constant_size > 0 ) {
-        constant_size_asr = make_ConstantWithType(make_IntegerConstant_t,
-                                constant_size, int_type, x->base.base.loc);
-        if( array_size == nullptr ) {
-            return constant_size_asr;
-        }
-    }
-    if( constant_size_asr ) {
-        array_size = builder.Add(array_size, constant_size_asr);
-    }
-
-    if( array_size == nullptr ) {
-        array_size = make_ConstantWithKind(make_IntegerConstant_t,
-            make_Integer_t, 0, 4, x->base.base.loc);
-    }
-    return array_size;
-}
-
 ASR::ttype_t* create_array_type_with_empty_dims(Allocator& al,
     size_t value_n_dims, ASR::ttype_t* value_type) {
     Vec<ASR::dimension_t> empty_dims; empty_dims.reserve(al, value_n_dims);
@@ -279,7 +109,7 @@ ASR::expr_t* create_temporary_variable_for_array(Allocator& al,
     // ArrayConstructor like [5, 2, 1]
     if (ASR::is_a<ASR::ArrayConstructor_t>(*value)) {
         ASR::ArrayConstructor_t* arr_constructor = ASR::down_cast<ASR::ArrayConstructor_t>(value);
-        value_m_dims->m_length = get_ArrayConstructor_size(al, arr_constructor);
+        value_m_dims->m_length = ASRUtils::get_ArrayConstructor_size(al, arr_constructor);
     }
     bool is_fixed_sized_array = ASRUtils::is_fixed_size_array(value_type);
     bool is_size_only_dependent_on_arguments = ASRUtils::is_dimension_dependent_only_on_arguments(
@@ -709,7 +539,7 @@ bool set_allocation_size(
             ASR::dimension_t allocate_dim;
             allocate_dim.loc = loc;
             allocate_dim.m_start = int32_one;
-            allocate_dim.m_length = get_ArrayConstructor_size(al,
+            allocate_dim.m_length = ASRUtils::get_ArrayConstructor_size(al,
                 ASR::down_cast<ASR::ArrayConstructor_t>(value));
             allocate_dims.push_back(al, allocate_dim);
             break;
@@ -719,7 +549,7 @@ bool set_allocation_size(
             ASR::dimension_t allocate_dim;
             allocate_dim.loc = loc;
             allocate_dim.m_start = int32_one;
-            allocate_dim.m_length = get_ArrayConstant_size(al,
+            allocate_dim.m_length = ASRUtils::get_ArrayConstant_size(al,
                 ASR::down_cast<ASR::ArrayConstant_t>(value));
             allocate_dims.push_back(al, allocate_dim);
             break;
@@ -950,7 +780,7 @@ class ArgSimplifier: public ASR::CallReplacerOnExpressionsVisitor<ArgSimplifier>
                 ASRUtils::ASRBuilder builder(al, loc);
                 ASR::dimension_t m_dim; m_dim.loc = loc;
                 m_dim.m_start = builder.i32(1);
-                m_dim.m_length = get_ImpliedDoLoop_size(al, implied_do_loop);
+                m_dim.m_length = ASRUtils::get_ImpliedDoLoop_size(al, implied_do_loop);
                 m_dims.push_back(al, m_dim);
                 ASR::ttype_t* type = ASRUtils::make_Array_t_util(al, loc,
                     implied_do_loop->m_type, m_dims.p, m_dims.size());
@@ -1119,7 +949,7 @@ class ArgSimplifier: public ASR::CallReplacerOnExpressionsVisitor<ArgSimplifier>
             ASRUtils::ASRBuilder builder(al, loc);
             ASR::dimension_t m_dim; m_dim.loc = loc;
             m_dim.m_start = builder.i32(1);
-            m_dim.m_length = get_ImpliedDoLoop_size(al, implied_do_loop);
+            m_dim.m_length = ASRUtils::get_ImpliedDoLoop_size(al, implied_do_loop);
             m_dims.push_back(al, m_dim);
             ASR::ttype_t* type = ASRUtils::make_Array_t_util(al, loc,
                 implied_do_loop->m_type, m_dims.p, m_dims.size());
